refactor: ditch layered architecture
