+++ README.md
@@ -17,9 +17,6 @@
 ## 💡 About
 anyguess will be a web app that allows you to create your own "geoguessr" style challenges with custom maps and landmarks from video games, for example
 
-## 👷‍♂️ Architecture
-I plan to develop this app using the **Layered Architecture** pattern. I was planning to use lucia-auth for authentication, but in the process of refactoring another app (which used next-auth and xata) to use drizzle, i managed to setup a fully working auth system using next-auth, drizzle-kit and xata in this repository (though ultimately i ditched xata). Next-auth is a great library which is very quick to setup at the cost of high levels of abstraction, so I'm not sure if it's the right fit for layered architecture, but I'm going to try it out.
-
 ## 📚 Key concepts
 - A single experience/custom level created by an user is internally referred to as a "Challenge", as to not cause confusion since in Javascript we have the `Map` class and `.map()` array method. In the interface, it's referred to as a "Map", which is how custom levels are referred to in the original geoguesser.
 - A "Challenge" is made up of the following:
